chore: reorganizes storage components
